refactor: 무한 스크롤 관련 코드 함수 이동 (임시)
